update DTO, conexiones base de datos

+++ src/main/java/pe/edu/upc/logisticmaster/backendandroid/backend/reserve/domain/model/ReserveCommand.java
@@ -1,14 +1,14 @@
 package pe.edu.upc.logisticmaster.backendandroid.backend.reserve.domain.model;
 
 public class ReserveCommand {
-    private Integer userId;
+    private Long userId;
     private Integer roomId;
     private String status;
     private String startTime;
     private String endTime;
 
     // Constructor
-    public ReserveCommand(Integer userId, Integer roomId, String status, String startTime, String endTime) {
+    public ReserveCommand(Long userId, Integer roomId, String status, String startTime, String endTime) {
         this.userId = userId;
         this.roomId = roomId;
         this.status = status;
@@ -17,11 +17,11 @@ public class ReserveCommand {
     }
 
     // Getters y Setters
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
